% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinkingGrid.R
\name{check_python_available}
\alias{check_python_available}
\title{Checks status of Python dependency.}
\usage{
check_python_available(install_if_NA = FALSE)
}
\arguments{
\item{install_if_NA}{{logical, optional} If TRUE, installs python 3.13 and required packages. Default is FALSE.}
}
\value{
None
}
\description{
Checks if python is available. Installs python if appropriate flag is set.
}
\examples{
\dontrun{
check_python_available()
check_python_available(install_if_NA = TRUE)
}

}
