% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils-2x2-plot.R
\name{create_test_2x2_plots}
\alias{create_test_2x2_plots}
\title{Creates an example list of four plots for use in 2x2 plot tests.}
\usage{
create_test_2x2_plots()
}
\value{
A list of four ggplot objects for testing purposes.
}
\description{
Creates four ggplot objects.
}
\examples{
# This is a list of four plots generated by some regression. Their exact nature is not
# important aside from the fact that they are ggplot objects.
plots <- ThinkingGrid::create_test_2x2_plots()

}
