% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tg.R
\name{create_tg_animation}
\alias{create_tg_animation}
\title{Illustration of create_tg_animation function}
\usage{
create_tg_animation(
  dc,
  ac,
  condition_col,
  type = "cells",
  proportion_type = "overall",
  filename = "tg_animation.gif",
  duration = 1,
  width = 800,
  height = 800,
  sorted_conditions = NULL,
  color_palette = "Greens",
  x_label = "Directedness",
  y_label = "Stickiness",
  max_legend = NULL,
  min_legend = NULL
)
}
\arguments{
\item{add_parameter_here}{{character} Description of the parameter}
}
\description{
Illustration of create_tg_animation function
}
\examples{
create_tg_animation(relevant_data)

}
