% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2x2.R
\name{default_inner_theme}
\alias{default_inner_theme}
\title{Default theme for subplots in a ThinkingGrid 2x2 quadrant plot.}
\usage{
default_inner_theme(inner_margin = 20)
}
\arguments{
\item{inner_margin}{{integer, optional} Controls the padding between the inner subplots.}
}
\value{
A theme object with the settings for the subplots of a 2x2 quadrant plot.
}
\description{
The default theme settings for the subplots of a 2x2 quadrant plot. In order to display
correctly, the important properties of the theme are:
\enumerate{
\item The background are set to transparant elements.
\item The main plot titles and legends are removed. (Axis labels are kept.)
\item The aspect ratio is set to 1.00.
\item Margins are adjusted to set the plot within the background squares.
}
}
\examples{
plots <- ThinkingGrid::create_test_2x2_plots()
p1 <- plots[[1]]
p2 <- plots[[2]]
p3 <- plots[[3]]
p4 <- plots[[4]]

# Here we can customize the theme.
itheme <- default_inner_theme(inner_margin = 10)

# This is the usual syntax.
thinkgrid_quadrant_plot(p1, p2, p3, p4, inner_theme = itheme)


}
