% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinkingGrid.R
\name{extract_quadrant_depths}
\alias{extract_quadrant_depths}
\title{Illustration of extract_quadrant_depths function}
\usage{
extract_quadrant_depths(
  data_file,
  dc_column = "Deliberate.Constraints",
  ac_column = "Automatic.Constraints"
)
}
\arguments{
\item{data_file}{{character, needed} Path to csv or excel file containing the data.}

\item{dc_column}{{character, optional} Name of the column containing deliberate constraints. Default is "Deliberate.Constraints".}

\item{ac_column}{{character, optional} Name of the column containing automatic constraints. Default is "Automatic.Constraints".}
}
\value{
data frame containing the quadrant depths. Columns include "Free.Depth", "Directed.Depth", "AffDir.Depth", and "Sticky.Depth".
}
\description{
Illustration of extract_quadrant_depths function
}
\details{
The function calculates the quadrant depths based on the deliberate and automatic constraints provided in the data file. The quadrant depths are calculated using the taxicab norm. Only one depth will be populated per observation, depending on the quadrant the observation falls into. The remaining three depths will be set to 0.
}
\examples{
# Calculate quadrant depths from survey data
data_file <- system.file("extdata", "sample_data.csv", package = "ThinkingGrid")
if (file.exists(data_file)) {
  depth_results <- extract_quadrant_depths(data_file, dc_column = "dc", ac_column = "ac")
}

}
