% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tg.R
\name{plot_tg}
\alias{plot_tg}
\title{Illustration of plot_tg function}
\usage{
plot_tg(
  dc,
  ac,
  proportion_type = "overall",
  type = "cells",
  color_palette = "Greens",
  x_label = "Directedness",
  y_label = "Stickiness",
  condition_col = NULL,
  comparison_type = "separate",
  pos_palette = "Greens",
  neg_palette = "Reds",
  max_legend = NULL,
  min_legend = NULL
)
}
\arguments{
\item{add_parameter_here}{{character} Description of the parameter}
}
\description{
Illustration of plot_tg function
}
\examples{
plot_tg(relevant_data)

}
