% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinkingGrid.R
\name{read_qualtrics_data}
\alias{read_qualtrics_data}
\title{Illustration of read_qualtrics_data function}
\usage{
read_qualtrics_data(data_file, setup_file)
}
\arguments{
\item{data_file}{{character, needed} The path to where the Qualtrics output is located. This should be a csv file that needs to be provided to this function UNEDITED.}

\item{setup_file}{{character, needed} setup_file used to generate the survey. This file should be the same file that was used to generate the survey. This file should have the same format as the survey_setup_file used in the generate_survey function.}
}
\value{
data frame containing the Qualtrics data. Columns include "uid", "Probe.Identifier", "Deliberate.Constraints", "Automatic.Constraints"

uid: Unique identifier for each participant. This correspons to the row number in the Qualtrics output file.

Probe.Identifier: Identifier for the probe. This is the same as the "id" column in the setup_file.

Deliberate.Constraints: The deliberate constraints provided by the participant. (X-axis on the thinking grid)

Automatic.Constraints: The automatic constraints provided by the participant. (Y-axis on the thinking grid)
}
\description{
Illustration of read_qualtrics_data function
}
\examples{
# Read Qualtrics survey data
setup_file <- system.file("extdata", "sample_setup_file.csv", package = "ThinkingGrid")
data_file <- system.file("extdata", "sample_qualtrics_output.csv", package = "ThinkingGrid")
if (file.exists(setup_file) && file.exists(data_file)) {
  survey_data <- read_qualtrics_data(data_file, setup_file)
}

}
