% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinkingGrid.R
\name{test_func}
\alias{test_func}
\title{Illustration of test_func function
#' A simple function to test the package.}
\usage{
test_func()
}
\description{
Illustration of test_func function
#' A simple function to test the package.
}
