% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2x2.R
\name{thinkgrid_quadrant_background}
\alias{thinkgrid_quadrant_background}
\title{Illustration of thinkgrid_quadrant_background function}
\usage{
thinkgrid_quadrant_background(
  arrowwidth = 1,
  title = NULL,
  xlab = "Executive Control",
  ylab = "Salience"
)
}
\description{
Creates a 6x6 grid background image for the 2x2 thinkgrid plots.
}
