% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2x2.R
\name{default_inner_theme}
\alias{default_inner_theme}
\title{Illustration of thinkgrid_quadrant_plot function}
\usage{
default_inner_theme(inner_margin = 20)
}
\arguments{
\item{inner_margin}{{integer, optional} Controls the padding between the inner subplots.}
}
\value{
A theme object with the settings for the subplots of a 2x2 quadrant plot.
}
\description{
The default theme settings for the subplots of a 2x2 quadrant plot. In order to display
correctly, the important properties of the theme are:
\enumerate{
\item The background are set to transparant elements.
\item The main plot titles and legends are removed. (Axis labels are kept.)
\item The aspect ratio is set to 1.00.
\item Margins are adjusted to set the plot within the background squares.
}
}
