% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinkingGrid.R
\name{install_thinkgrid}
\alias{install_thinkgrid}
\title{Illustration of install_thinkgrid function}
\usage{
install_thinkgrid(envname = "r-thinkgrid")
}
\arguments{
\item{envname}{{character, optional} Name of the virtual environment to be created. Default is "r-thinkgrid".}
}
\value{
None
}
\description{
Creates a virtual environment and installs the required python packages
}
\examples{
install_thinkgrid()
install_thinkgrid("new_environment_name")

}
