% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2x2.R
\name{thinkgrid_quadrant_background}
\alias{thinkgrid_quadrant_background}
\title{Illustration of thinkgrid_quadrant_background function}
\usage{
thinkgrid_quadrant_background()
}
\description{
Does something.
}
