% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2x2.R
\name{thinkgrid_quadrant_plot}
\alias{thinkgrid_quadrant_plot}
\title{Illustration of thinkgrid_quadrant_plot function}
\usage{
thinkgrid_quadrant_plot(
  p_sticky,
  p_salience,
  p_free,
  p_directed,
  inner_theme = NULL,
  inner_margin = 20,
  arrowwidth = 1,
  xlab = "Executive Control",
  ylab = "Salience"
)
}
\arguments{
\item{p_sticky}{{ggplot} A ggplot object for the "Sticky" quadrant.}

\item{p_salience}{{ggplot} A ggplot object for the "Salience" quadrant.}

\item{p_free}{{ggplot} A ggplot object for the "Free" quadrant.}

\item{p_directed}{{ggplot} A ggplot object for the "Directed" quadrant.

#' @return A list containing two grid objects: one with the quadrant background and the plots, and another with the legend.}
}
\description{
Creates a 2x2 quadrant plot with four ggplot objects.
}
