% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2x2.R
\name{thinkgrid_quadrant_plot}
\alias{thinkgrid_quadrant_plot}
\title{Illustration of thinkgrid_quadrant_plot function}
\usage{
thinkgrid_quadrant_plot(
  p_sticky,
  p_salience,
  p_free,
  p_directed,
  inner_theme = NULL,
  arrowwidth = 1,
  xlab = "Executive Control",
  ylab = "Salience"
)
}
\arguments{
\item{p_sticky}{(ggplot or rastergrob) A ggplot object for the "Sticky" quadrant.}

\item{p_salience}{(ggplot or rastergrob) A ggplot object for the "Salience" quadrant.}

\item{p_free}{(ggplot or rastergrob) A ggplot object for the "Free" quadrant.}

\item{p_directed}{(ggplot or rastergrob) A ggplot object for the "Directed" quadrant.}

\item{inner_theme}{(theme, optional) A theme for the inner subplots.
See \code{default_inner_theme} for more details.}

\item{arrowwidth}{(integer, optional) Controls the thickness of the axis arrows.}

\item{xlab}{(character, optional) Label for the x-axis.}

\item{ylab}{(character, optional) Label for the y-axis.}
}
\value{
A ggplot object (created via cowplot) which consists of the thinking grid
background and the inlayed 2x2 subplots (or images) corresponding to the respective quadrant.
}
\description{
Creates a 2x2 quadrant plot with four ggplot objects.
}
