% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinkingGrid-package.R
\docType{package}
\name{ThinkingGrid-package}
\alias{ThinkingGrid}
\alias{ThinkingGrid-package}
\title{ThinkingGrid: What the Package Does (One Line, Title Case)}
\description{
A description here.
}
\author{
\strong{Maintainer}: Vishal Kuvar \email{kuvar001@umn.edu} (\href{https://orcid.org/0000-0002-6799-9790}{ORCID})

}
\keyword{internal}
